Songs.js Clickable List
Got working method to display lead sheets along with the tracks. List is clickable, and
able to switch between songs. The filepath to the lead sheet and backing track url are stored in the state
of the songList component. Tried to make songList have child components songs, but I couldn't get it to work.
Kept the name the same because songList is more relevant.