Song Display changes + CSS
Made some changes to the song display page, made how the resources are displayed more dynamic and added some simple CSS to the header
Also implemented a rough timestamping feature to track the amount of time spent on each song. Time spent on each tune is shown as an alert.